Change how colinearity is fixed
There's a risk of colinearity between the covariates "p_income", "p_age_group", and "employ": p_income and employ for all (and onlyfor ) individuals below 18 is set to "Child - Not Applicable". Therefore, in the previous version, the model could not distinguish between income-child, employ-child, and age_group-0-4.. This commit fixes this problem.